backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
PyMuPDF==1.23.26
pdf2image==1.17.0
pytesseract==0.3.10
psycopg2-binary==2.9.9
python-dotenv==1.0.1 